requirements.txt
sortedcontainers
munch
dnspython
tqdm
pycrypto
numpy
pyopencl
